Refactor: MyEats 화면 및 즐겨찾기 화면 UI 컴포저블 분리 및 네비게이션 로직 추가
- `MyEatsScreen.kt`: `MyEatsNums`와 `FavoriteRow` 컴포저블을 별도의 파일로 분리하여 코드 가독성을 개선했습니다.
- `FavoriteScreen.kt`:
    - `FavoriteTopBar` 컴포저블을 별도의 파일로 분리했습니다.
    - 즐겨찾기 목록 아이템 클릭 시 즐겨찾기 상세 화면(`FavoriteDetailScreen`)으로 이동하는 네비게이션 로직을 추가했습니다. (`onNavigateToDetailPage` 콜백 사용)
- `FavoriteDetailScreen.kt`: 즐겨찾기 상세 화면을 위한 기본 UI 구조(`FavoriteDetailScreen`, `FavoriteDetailTopBar`)를 추가했습니다. (상세 구현은 미완료)
- `MainNavHost.kt`: `FavoriteScreen`에서 `FavoriteDetailScreen`으로의 네비게이션을 정의하고, `RestaurantData`를 전달하도록 수정했습니다.
- `MainActivity.kt`: 현재 화면이 `FavoriteScreen`일 경우 하단 네비게이션 바를 숨기도록 로직을 추가했습니다.
- `RestaurantData.kt`: `RestaurantData` 클래스에 `@Serializable` 어노테이션을 추가하여 객체 직렬화/역직렬화를 가능하게 했습니다.
- `Route.kt`: `FavoriteDetail` 라우트를 추가하여 즐겨찾기 상세 화면으로 이동 시 `RestaurantData`를 전달할 수 있도록 했습니다.
