+++ app/src/main/java/com/kuit/kuit6android/navigation/Route.kt
@@ -1,5 +1,6 @@
 package com.kuit.kuit6android.navigation
 
+import com.kuit.kuit6android.ui.home.data.RestaurantData
 import kotlinx.serialization.Serializable
 
 sealed interface Route {
@@ -19,6 +20,11 @@ sealed interface Route {
 
     @Serializable
     data object Favorite : Route
+
+    @Serializable
+    data class FavoriteDetail(
+        val restaurantData: RestaurantData,
+    ) : Route
 
     @Serializable
     data object OrderHistory : Route
