+++ app/src/main/java/com/kuit/kuit6android/ui/myeats/screen/MyEatsScreen.kt
@@ -1,31 +1,24 @@
 package com.kuit.kuit6android.ui.myeats.screen
 
 import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.kuit.kuit6android.R
 import com.kuit.kuit6android.navigation.Route
 
 @Composable
@@ -82,36 +75,7 @@ fun MyEatsScreen(
             )
         )
 
-        Row(
-            modifier = modifier
-                .fillMaxWidth()
-                .padding(horizontal = 50.dp),
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            val myEatsNums = mapOf(
-                "내가 남긴 리뷰" to 0,
-                "도움이 됐어요" to 0,
-                "즐겨찾기" to 0
-            )
-            myEatsNums.forEach {
-                Column(
-                    modifier = Modifier.weight(weight = 1f),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.spacedBy(space = 14.dp)
-                ) {
-                    Text(
-                        text = "${it.value}",
-                        fontWeight = FontWeight.SemiBold,
-                        fontSize = 28.sp
-                    )
-                    Text(
-                        text = it.key,
-                        fontSize = 11.sp,
-                        fontWeight = FontWeight.Normal
-                    )
-                }
-            }
-        }
+        MyEatsNums(modifier)
 
         Spacer(
             modifier = Modifier.height(
@@ -143,51 +107,13 @@ fun MyEatsScreen(
                 fontWeight = FontWeight.Bold
             )
         }
+
         Spacer(
             modifier = Modifier.height(
                 height = 22.dp
             )
         )
-        Row(
-            modifier = modifier
-                .fillMaxWidth()
-                .padding(
-                    start = 32.dp,
-                    end = 211.dp
-                )
-                .clickable(
-                    onClick = {
-                        onNavigateToFavorite(
-                            Route.Favorite
-                        )
-                    }
-                ),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                painter = painterResource(
-                    id = R.drawable.like
-                ),
-                contentDescription = "즐겨찾기",
-                modifier = modifier.size(
-                    width = 19.98.dp,
-                    height = 18.01.dp
-                )
-            )
-            Text(
-                text = "즐겨찾기",
-                fontSize = 21.sp,
-                fontWeight = FontWeight.Normal
-            )
-        }
+
+        FavoriteRow(modifier, onNavigateToFavorite)
     }
 }
-//
-//@Preview(showBackground = true)
-//@Composable
-//private fun MyEatsScreenPreview() {
-//    MyEatsScreen(
-//        padding = PaddingValues()
-//    )
-//}
